Set up notices for AGPL compliance
Serve an HTML copy of the AGPL from the site.
Rewrite README with licensing info.
Change footer to include copyright notice for templates.

Still need footers to acknowledge content copyright.  I'd also like to
set up badges for Creative Commons licensing, and make that appear in
the paginated index of the blog.